Solved bug of not working second time
Before, there was an issue that, after the first = operation, the second = wouldn't work. This was due to '=' being saved as the ongoing operator. I added a protective if-statement that prevents '=' from being the ongoing operator. This way the desired behavior of 2 * 3 * 4 showing "6" on the display after pressing the second '*'.